refactor(v2): rename task processor occurrences to aggregator processor (#722)
