+++ aggregator/aggregator.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"math/big"
 
-	taskprocessor "github.com/Layr-Labs/eigensdk-go/aggregator/task-processor"
 	"github.com/Layr-Labs/eigensdk-go/logging"
 	taskmanager "github.com/Layr-Labs/eigensdk-go/task-manager"
 	sdktypes "github.com/Layr-Labs/eigensdk-go/types"
@@ -27,7 +26,7 @@ import (
 //   - Receiving signed responses from the operators.
 //   - Sending the aggregated responses to the `TaskManager` contract
 //
-// Most of these things are delegated to the `TaskProcessor` interface, that processes
+// Most of these things are delegated to the `Processor` interface, that processes
 // tasks and communicates with the on-chain `TaskManager` contract.
 type Aggregator[Input any, Output any] struct {
 	logger logging.Logger
@@ -44,15 +43,15 @@ type Aggregator[Input any, Output any] struct {
 	// ABI of the task manager contract
 	taskManagerAbi *abi.ABI
 
-	taskProcessor taskprocessor.TaskProcessor[Input, Output]
+	processor Processor[Input, Output]
 }
 
-// NewAggregator creates a new Aggregator with the provided config, a logger, a task processor and the task manager contract's ABI.
+// NewAggregator creates a new Aggregator with the provided config, a logger, a processor and the task manager contract's ABI.
 func NewAggregator[Input any, Output any](
 	logger logging.Logger,
 	c Config,
 	taskManagerAbi *abi.ABI,
-	taskProcessor taskprocessor.TaskProcessor[Input, Output],
+	processor Processor[Input, Output],
 ) (*Aggregator[Input, Output], error) {
 	avsRegistryConfig := avsregistry.Config{
 		RegistryCoordinatorAddress:    c.RegistryCoordinatorAddress,
@@ -94,7 +93,7 @@ func NewAggregator[Input any, Output any](
 			logger.Error("task Response could not be converted to sdk aggregator's Task Response type")
 		}
 
-		return taskProcessor.ProcessTaskResponse(taskResponse)
+		return processor.ProcessTaskResponse(taskResponse)
 	}
 
 	avsRegistryService := avsregistryservice.NewAvsRegistryServiceChainCaller(avsRegistryReader, operatorPubkeysService, logger)
@@ -118,7 +117,7 @@ func NewAggregator[Input any, Output any](
 		blsAggregationService: blsAggregationService,
 		newTaskCreatedLogs:    newTaskCreatedLogs,
 		taskManagerAbi:        taskManagerAbi,
-		taskProcessor:         taskProcessor,
+		processor:             processor,
 	}, nil
 }
 
@@ -174,7 +173,7 @@ func (agg *Aggregator[Input, Output]) run(ctx context.Context) error {
 }
 
 // When processing a new task event, the aggregator unpacks the log data into the new task created event and
-// sends it to the task processor
+// sends it to the processor
 func (agg *Aggregator[Input, Output]) processNewTask(log types.Log) (blsagg.TaskMetadata, error) {
 	var newTaskCreatedLog taskmanager.NewTaskCreatedEvent[Input]
 
@@ -190,7 +189,7 @@ func (agg *Aggregator[Input, Output]) processNewTask(log types.Log) (blsagg.Task
 
 	newTask := newTaskCreatedLog.Task
 
-	metadata, err := agg.taskProcessor.ProcessNewTask(newTaskIndex, newTask)
+	metadata, err := agg.processor.ProcessNewTask(newTaskIndex, newTask)
 	if err != nil {
 		return blsagg.TaskMetadata{}, err
 	}
@@ -198,7 +197,7 @@ func (agg *Aggregator[Input, Output]) processNewTask(log types.Log) (blsagg.Task
 	return metadata, nil
 }
 
-// When processing an aggregated response, the aggregator delegates the processing to the task processor
+// When processing an aggregated response, the aggregator delegates the processing to the processor
 func (agg *Aggregator[Input, Output]) processAggregatedResponse(
 	response blsagg.BlsAggregationServiceResponse,
 ) error {
@@ -230,7 +229,7 @@ func (agg *Aggregator[Input, Output]) processAggregatedResponse(
 		agg.logger.Error("task Response could not be converted to sdk aggregator's Task Response type")
 	}
 
-	err := agg.taskProcessor.ProcessAggregatedResponse(response.TaskIndex, taskResponse, nonSignerStakesAndSignature)
+	err := agg.processor.ProcessAggregatedResponse(response.TaskIndex, taskResponse, nonSignerStakesAndSignature)
 	if err != nil {
 		return utils.WrapError("Aggregator failed to respond to task", err)
 	}
