+++ aggregator/processor.go
@@ -1,4 +1,4 @@
-package taskprocessor
+package aggregator
 
 import (
 	blsagg "github.com/Layr-Labs/eigensdk-go/services/bls_aggregation"
@@ -6,10 +6,11 @@ import (
 	sdktypes "github.com/Layr-Labs/eigensdk-go/types"
 )
 
-// The task processor processes new tasks, task responses, and aggregated responses.
-// A default implementation for this trait can be the Indexing Task Processor, that can be found in
-// aggregator/task-processor/indexing_task_processor.go
-type TaskProcessor[Input any, Output any] interface {
+// The processor processes new tasks, task responses, and aggregated responses.
+// A default implementation for this trait can be the Indexing Processor, that can be found in
+// aggregator/indexing_processor.go
+// Note: Do not confuse it with the challenger module's Processor interface
+type Processor[Input any, Output any] interface {
 	// Processes a new task, and returns the task metadata to be sent to the BLS aggregation service, or an error
 	ProcessNewTask(taskIndex sdktypes.TaskIndex, task taskmanager.Task[Input]) (blsagg.TaskMetadata, error)
