+++ challenger/challenger.go
@@ -16,6 +16,7 @@ import (
 )
 
 // The Challenger processor is responsible for processing the challenges
+// Note: Do not confuse it with the aggregator module's Processor interface
 type ChallengerProcessor[Input any, Output any] interface {
 	// Processes new tasks, returns an error in case of failure
 	ProcessNewTaskCreated(taskIndex uint32, task taskmanager.Task[Input]) error
